Fix shoot command and robotInit bug
We fixed the reason for why the code couldn’t been seen in the DS even
when everything built normally. The problem was that the OI was
initialized before the shooter subsystem. The OI contains a command in
the constructor that uses the shooter subsystem. Since the subsystem
does not “exist” when the OI is instantiated, it gives a NPE. We also
fixed the shoot command itself to allow control with the trigger on a
joystick.